枚举类型迁移至新命名空间并类型安全升级
本次提交将所有枚举及DTO类型统一迁移至 NewLife.AI.ChatAI 命名空间，数据模型相关字段由 Int32/String 改为枚举类型，Model.xml 同步更新。高级查询方法、控制器、服务、测试等处均适配枚举类型参数，提升类型安全。McpServerConfig 的 TransportType 改为枚举，接口参数同步调整。删除 ChatMessage 的更新相关字段。整体提升代码规范性和可维护性。

+++ XUnitTest/InMemoryChatApplicationService.cs
@@ -12,8 +12,8 @@ using Int64 = System.Int64;
 using String = System.String;
 using Boolean = System.Boolean;
 using CancellationToken = System.Threading.CancellationToken;
-using NewLife.AI.ChatAI.Contracts;
 using NewLife.AI.Models;
+using NewLife.AI.ChatAI;
 
 namespace NewLife.ChatAI.Services;
 
@@ -147,7 +147,7 @@ public class InMemoryChatApplicationService
 
         var modelCode = _conversations.TryGetValue(source.ConversationId, out var conv) ? conv.ModelCode : "qwen-max";
         var assistantId = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds();
-        yield return ChatStreamEvent.MessageStart(assistantId, modelCode, (Int32)source.ThinkingMode);
+        yield return ChatStreamEvent.MessageStart(assistantId, modelCode, source.ThinkingMode);
 
         var answer = "这是编辑后重新生成的流式回复。";
         var content = new StringBuilder();
@@ -194,7 +194,7 @@ public class InMemoryChatApplicationService
         }
 
         var modelCode = _conversations.TryGetValue(source.ConversationId, out var conv) ? conv.ModelCode : "qwen-max";
-        yield return ChatStreamEvent.MessageStart(source.Id, modelCode, (Int32)source.ThinkingMode);
+        yield return ChatStreamEvent.MessageStart(source.Id, modelCode, source.ThinkingMode);
 
         var answer = "这是重新生成的流式回复。";
         var content = new StringBuilder();
@@ -238,7 +238,7 @@ public class InMemoryChatApplicationService
 
         // message_start（含模型和思考模式）
         var modelCode = _conversations.TryGetValue(conversationId, out var conv) ? conv.ModelCode : "qwen-max";
-        yield return ChatStreamEvent.MessageStart(assistantMessageId, modelCode, (Int32)request.ThinkingMode);
+        yield return ChatStreamEvent.MessageStart(assistantMessageId, modelCode, request.ThinkingMode);
 
         var answer = "这是流式回复骨架。后续可接入真实模型推理与上下文管理。";
         var chunks = answer.Split('。', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
